requirements.txt
# Ergotech — Dépendances Streamlit Cloud

anthropic>=0.40.0
langchain-anthropic>=0.3.0
langchain>=0.3.0
langchain-community>=0.3.0
langchain-core>=0.3.0
chromadb>=0.5.0
pymupdf>=1.24.0
streamlit>=1.40.0
python-dotenv>=1.0.0
tiktoken>=0.7.0
pydantic>=2.0.0
